+++ loudness-scanner/loudness-scanner-source.sh
@@ -33,6 +33,10 @@ then
 fi
 
 git submodule update --depth=1 --init --recursive --progress
+if [ $? -ne 0 ]; then
+    echo "Problem with submodules"
+    exit 1
+fi
 find . -name .git -exec rm -rf {} \;
 cd ..
 tar cvfz loudness-scanner.tar.gz loudness-scanner/*
